Create transducer function file
